Release v0.9.4: Bug fixes and GUI improvements
Unified Fit GUI
- Copy/Swap button capped to 120px (same as Graph Unified)
- Control panel now resizable (removed Fixed size policy / max-width lock)
- Parameter value fields narrowed 10%, limit fields 5% to save space
- Fitting max_iterations raised 1000 → 5000 to reduce premature convergence

Batch / scripting — _pyirena_config round-trip
- fit_unified now embeds _pyirena_config in saved HDF5 (fitted values)
- fit_sizes, fit_modeling, fit_simple, fit_waxs_peaks all pass setup_state
- HDF5 num_levels/level_number attributes stored as int32 for Igor compat

Save Params to JSON (all tools)
- Single dialog replaces two: OS "replace file" prompt suppressed via
  DontConfirmOverwrite | DontUseNativeDialog; one clear message shown
- Message accurately states only the tool section is updated, not the file
- Modeling export now merges into existing file instead of replacing it
- WAXS Peak Fit export adds the same section-exists check as other tools

Data Merge batch
- Silent fail fixed: _load_file suppresses modal dialog in batch (quiet=True);
  entire per-pair block caught by BaseException (re-raises Ctrl-C/SystemExit)
- Q overlap check uses finite-only max/min to handle NaN edge cases
- Live progress shown in status bar: "Merging N/total: filename …"
- Scale fitting bounds widened 0.01–100 → 0.001–1000

Sizes distribution panel
- "Set Q from cursors" button enlarged and moved inline with Q min/Q max
- "Q range for fit:" label centred in both Power-Law and Flat Background sections
- Power-law no-checkbox error now also highlighted in control-panel status bar
- Number of bins max raised 500 → 501; silent snap: multiples of 100 → +1

Igor import (h5xp)
- WAVE_PICKERS_H5XP["USAXS"] now recognises DSM_* (desmeared) and SMR_*
  (slit-smeared) wave names from Igor-exported h5xp, matching pxp behaviour
- SAXS/WAXS h5xp picker adds R_Qvec/R_Int/R_Error for Igor-exported files

Data Merge state persistence
- Folder paths (DS1, DS2, output) now survive app restart via StateManager
- hideEvent added so Cmd+W on macOS also triggers save_state
- DS1 pre-populate only runs on first-ever open when no saved state exists

Miscellaneous
- About dialog version now reads from pyirena.__version__ (no manual update)
- Level/Peak selector in HDF5 Viewer Collect tab widened to match Item combo
- Data Merge batch guard skips and reports bad pairs instead of halting

+++ pyirena/io/nxcansas_unified.py
@@ -250,7 +250,7 @@ def save_unified_fit_results(filepath: Path,
         unified_group.attrs['analysis_type'] = 'Unified Fit'
         unified_group.attrs['program'] = 'pyirena'
         unified_group.attrs['timestamp'] = timestamp
-        unified_group.attrs['num_levels'] = num_levels
+        unified_group.attrs['num_levels'] = np.int32(num_levels)
         # Store fit results as scalar datasets (browseable/collectable in HDF5 viewer)
         unified_group.create_dataset('background', data=float(background))
         unified_group.create_dataset('chi_squared', data=float(chi_squared))
@@ -287,7 +287,7 @@ def save_unified_fit_results(filepath: Path,
         for i, level_params in enumerate(levels):
             level_num = i + 1
             level_group = unified_group.create_group(f'level_{level_num}')
-            level_group.attrs['level_number'] = level_num
+            level_group.attrs['level_number'] = np.int32(level_num)
 
             for param_name, param_value in level_params.items():
                 if isinstance(param_value, (int, float)) and not isinstance(param_value, bool):
